+++ Source/DafnyTestGeneration/Utils.cs
@@ -1,6 +1,8 @@
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using DafnyServer.CounterexampleGeneration;
+using Microsoft.Boogie;
 using Microsoft.Dafny;
 using Errors = Microsoft.Dafny.Errors;
 using Parser = Microsoft.Dafny.Parser;
@@ -24,8 +26,25 @@ namespace DafnyTestGeneration {
       if (success) {
         program = new Program(fileName, module, builtIns, reporter);
       }
-      new Resolver(program).ResolveProgram(program);
       return program;
+    }
+
+    /// <summary>
+    /// Deep clone a Boogie program.
+    /// </summary>
+    public static Microsoft.Boogie.Program
+      DeepCloneProgram(Microsoft.Boogie.Program program) {
+      var oldPrintInstrumented = DafnyOptions.O.PrintInstrumented;
+      var oldPrintFile = DafnyOptions.O.PrintFile;
+      DafnyOptions.O.PrintInstrumented = true;
+      DafnyOptions.O.PrintFile = "-";
+      var output = new StringWriter();
+      program.Emit(new TokenTextWriter(output, DafnyOptions.O));
+      var textRepresentation = output.ToString();
+      Microsoft.Boogie.Parser.Parse(textRepresentation, "", out var copy);
+      DafnyOptions.O.PrintInstrumented = oldPrintInstrumented;
+      DafnyOptions.O.PrintFile = oldPrintFile;
+      return copy;
     }
   }
 }